Location tracking, display location details and show journey detail implemented.

+++ app/src/main/java/com/geekmk/mtracker/helper/AppPreferences.java
@@ -16,7 +16,7 @@ public class AppPreferences {
     return mPrefs.getLong(context.getString(R.string.val_current_journey), 0L);
   }
 
-  public static void setJourneyId(int uniqueId, Context context) {
+  public static void setJourneyId(long uniqueId, Context context) {
     SharedPreferences.Editor mPrefs = context
         .getSharedPreferences(context.getString(R.string.prefs), Context.MODE_PRIVATE).edit();
     mPrefs.putLong(context.getString(R.string.val_current_journey), uniqueId);
